MediaImpl: remove gst_sample_ref to avoid leak
The gst_sample pulled from the appsink is already reference counted. We just
need to unref it after copying it, which we do.

If we reference it again, the sample/buffer is not freed - I have seen the app
crashing while consuming most of system memory (can lead OS to out-of-memory
condition) before this commit.